Move extensions to new namespace
