+++ src/AudioBand/ViewModels/PlaybackControlsAppearance.cs
@@ -5,6 +5,7 @@ using System.ComponentModel;
 using System.Drawing;
 using System.IO;
 using System.Runtime.CompilerServices;
+using AudioBand.Extensions;
 using NLog;
 
 namespace AudioBand.ViewModels
